Create a separate module for creating the statement data
